requirements.txt
flaskFlask-SQLAlchemypsycopg2-binarygunicorn